rename 'parser' plugins to 'transformer'
