+++ examples/gatsbygram/package.json
@@ -5,18 +5,18 @@
   "version": "1.0.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
-    "gatsby": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-link": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-parser-json": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-parser-sharp": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-glamor": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-google-analytics": "^1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-manifest": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-offline": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-sharp": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-source-filesystem": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-filesystem": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-sharp": "1.0.0-alpha12-alpha.1fdb9004",
+    "gatsby": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-link": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-transformer-json": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-transformer-sharp": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-glamor": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-google-analytics": "^1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-manifest": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-offline": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-sharp": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-source-filesystem": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-filesystem": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-sharp": "1.0.0-alpha12-alpha.51e1923f",
     "instagram-screen-scrape": "^2.0.0",
     "lodash": "^4.16.4",
     "mkdirp": "^0.5.1",
