+++ packages/gatsby-transformer-json/.gitignore
