+++ packages/gatsby-transformer-json/.npmignore
