+++ packages/gatsby-transformer-json/index.js
