+++ packages/gatsby-transformer-remark/.npmignore
