+++ packages/gatsby-transformer-remark/index.js
