+++ packages/gatsby-transformer-sharp/.gitignore
