+++ packages/gatsby-transformer-sharp/.npmignore
