+++ packages/gatsby-transformer-sharp/index.js
