+++ packages/gatsby-transformer-sharp/src/gatsby-node.js
