+++ packages/gatsby-transformer-yaml/.npmignore
