+++ packages/gatsby-transformer-yaml/index.js
