+++ packages/gatsby-transformer-yaml/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "gatsby-parser-sharp",
+  "name": "gatsby-transformer-yaml",
   "version": "1.0.0-alpha12",
-  "description": "Gatsby parser plugin for images using Sharp",
+  "description": "Gatsby transformer plugin for yaml",
   "scripts": {
     "build": "babel src --out-dir .",
     "watch": "babel -w src --out-dir ."
@@ -9,13 +9,13 @@
   "keywords": [
     "gatsby",
     "gatsby-plugin",
-    "sharp",
-    "image"
+    "yaml"
   ],
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "license": "MIT",
   "dependencies": {
     "bluebird": "^3.4.6",
+    "js-yaml": "3.7.0",
     "unist-util-select": "^1.5.0"
   },
   "devDependencies": {
