+++ packages/gatsby/lib/schema/build-node-types.js
@@ -132,7 +132,6 @@ module.exports = async () =>
                 name: _.camelCase(`${typeName} field`),
                 type: nodeType.nodeObjectType,
                 resolve: (node, a, context, { fieldName }) => {
-                  const mapping = store.getState().config.mapping
                   const fieldSelector = `${node.___path}.${fieldName}`
                   let fieldValue = node[fieldName]
                   const sourceFileNode = _.find(
