+++ www/package.json
@@ -7,27 +7,27 @@
   "dependencies": {
     "bluebird": "^3.4.6",
     "fs-extra": "^2.0.0",
-    "gatsby": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-link": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-parser-remark": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-parser-sharp": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-parser-yaml": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-catch-links": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-glamor": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-google-analytics": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-manifest": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-offline": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-plugin-sharp": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-source-filesystem": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-filesystem": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-remark": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-remark-autolink-headers": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-remark-copy-linked-files": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-remark-prismjs": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-remark-responsive-iframe": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-remark-responsive-image": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-remark-smartypants": "1.0.0-alpha12-alpha.1fdb9004",
-    "gatsby-typegen-sharp": "1.0.0-alpha12-alpha.1fdb9004",
+    "gatsby": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-link": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-transformer-remark": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-transformer-sharp": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-transformer-yaml": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-catch-links": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-glamor": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-google-analytics": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-manifest": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-offline": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-plugin-sharp": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-source-filesystem": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-filesystem": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-remark": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-remark-autolink-headers": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-remark-copy-linked-files": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-remark-prismjs": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-remark-responsive-iframe": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-remark-responsive-image": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-remark-smartypants": "1.0.0-alpha12-alpha.51e1923f",
+    "gatsby-typegen-sharp": "1.0.0-alpha12-alpha.51e1923f",
     "lodash": "^4.16.6",
     "parse-filepath": "^1.0.1",
     "react-helmet": "^4.0.0",
